Word2Vec Tutorial - The Skip-Gram Model
The author of the blog took an attempt to discuss the skip gram neural network architecture
for Word2Vec. The model described in the blog, uses a simple neural network with a single
hidden layer which is used to learn the weights of the hidden layer. The goal of the learned
weights is to output a set of nearby words relating to the input word. This Word2Vec model
transforms word to its vector representation and also reserves the similarity between related
words. It is done by back propagation of the generated output and the expected output
of the model. This model solves one of the basic properties of natural language which is
the context of the word in a sentence. Word2Vec model solves this problem by generating
similar vectors for similar words. For example, the word “cat” and “dog” will have small
cosine distance of their vectors.
One of the major accomplishments of this model is that the paper [1] introducing the
model, works with a very large corpus. This large amount of data ensures high-quality word
vectors. Another notable accomplishment of this Word2Vec model is that we can establish
algebraic relations between similar words. For example, the vector(“King”) - vector(“Man”)
+ vector(“Woman”) results in a vector that is closest to the vector representation of the
word “Queen”.